refactor: decouple server invoke from register
